refactor(tools): lift the sandbox wrapping and the read-only kubectl policy into infra/
Both were exported from `cmd-exec/restricted-bash.ts`, so anything else that needed them had to
import a TOOL module. Three `infra/` tests were already doing exactly that — `kubectl-grammar.test.ts`
and `command-sets.test.ts` import `validateKubectlInPipeline` from the tool file, and `kubectl.test.ts`
names it in prose — which is the same inversion one layer up.

Three modules now: `infra/sandbox-exec.ts` (the `sudo -u sandbox` + `timeout -k` command line and the
session reap), `infra/kubectl-readonly-policy.ts` (`validateKubectlInPipeline` and the three helpers
it owns), and `infra/bounded-exec.ts` (moved from `cmd-exec/`, where it was infrastructure sitting
among tools). `restricted-bash.ts` re-exports all of them, so no call site changes.

A pure move: every relocated body is byte-identical to its predecessor, checked mechanically against
the parent commit rather than by eye. The one edit is `isDiagnosticApiPath`'s parameter, renamed
`path` -> `apiPath` so it cannot shadow a `node:path` import in its new home.

Two test changes come with it, both because the move breaks them:

`restricted-bash-fallback.test.ts` mocks bounded-exec BY PATH, so moving the module would have left
the mock pointing at a file that no longer exists — silently intercepting nothing while the test kept
passing for the wrong reason.

`credential-isolation-invariants.test.ts` asserted the sandbox drop lived in restricted-bash.ts and
in none of six named files. Rather than repoint the path, it now enumerates every tracked source file
and requires exactly one to contain the drop, which checks the "one code path" claim directly instead
of via a deny-list covering only the files someone remembered. Doing that closed a blind spot it
already had: its regex matched only the shell-string spelling, so `spawn("sudo", ["-n","-E","-u",
"sandbox", …])` — how the session reap becomes sandbox — would have passed in any file at all. Both
spellings are matched now, and both were confirmed to fail the test when a second drop is introduced.
